Subclass the editor's BlockTag so we can pick up some heuristic changes they are adding there.
